+++ apix/explore.py
@@ -10,7 +10,7 @@ import time
 import yaml
 from logzero import logger
 from pathlib import Path
-from apix.parsers import apipie
+from apix.parsers import apipie, test
 
 
 
@@ -25,21 +25,26 @@ class AsyncExplorer():
     _queue = attr.ib(default=[], repr=False)
 
     def __attrs_post_init__(self):
+        """perform the more complex steps of class initialization"""
         if not self.version:
             self.version = time.strftime('%Y-%m-%d', time.localtime())
         # choose the correct parser class from known parsers
         if self.parser.lower() == 'apipie':
             self.parser = apipie.APIPie()
+        elif self.parser.lower() == 'test':
+            self.parser = test.TestParser()
         if not self.parser or isinstance(self.parser, str):
             logger.warning('No known parser specified! Please review documentation.')
 
     async def _async_get(self, session, link):
+        """visit a page and download the content, returning the link and content"""
         async with session.get(self.host_url + link[1], verify_ssl=False) as response:
             content = await response.read()
             logger.debug(link[1])
             return (link, content)
 
     async def _async_loop(self, links):
+        """asynchronously visit each stored link and store results"""
         tasks = []
         async with aiohttp.ClientSession() as session:
             for link in links:
@@ -51,6 +56,7 @@ class AsyncExplorer():
                 self._queue.append(result)
 
     def _visit_links(self, links, retries=3):
+        """main controller for asynchronous page visiting, will attempt 3 retries"""
         try:
             loop = asyncio.get_event_loop()
             loop.run_until_complete(self._async_loop(links))
@@ -63,12 +69,14 @@ class AsyncExplorer():
                 self._visit_links(links, retries - 1)
 
     def _link_params(self):
+        """scrape the content downloaded from each page"""
         while self._queue:
             link, content = self._queue.pop(0)
             logger.debug(f'Scraping {link[1]}')
             self._data[link[1]] = self.parser.scrape_content(content)
 
     def save_data(self):
+        """convert the stored data into yaml-friendly dict and save"""
         yaml_data = self.parser.yaml_format(self._data)
         if not yaml_data:
             logger.warning('No data to be saved. Exiting.')
@@ -84,11 +92,15 @@ class AsyncExplorer():
         logger.info(f'Saving results to {fpath}')
         with fpath.open('w+') as outfile:
             yaml.dump(yaml_data, outfile, default_flow_style=False)
+        return fpath
 
     def explore(self):
-        if 'apidoc/' not in self.base_path:
-            logger.warning('I don\'t know how to explore that yet.')
-            return
+        """main function for the explore module
+        visit the initial page, pulling all links from the page
+        grab content at link found
+        sort the links to account for incorrect async return order
+        scrape the content at each page, per the parser's direction
+        """
         result = requests.get(self.host_url + self.base_path, verify=False)
         if not result:
             logger.warning(f"I couldn't find anything useful at "
@@ -102,3 +114,4 @@ class AsyncExplorer():
         # sort the results by link name, to normalize return order
         self._queue = sorted(self._queue, key=lambda x: x[0][1])
         self._link_params()
+        return True